agent/requirements.txt
requests==2.31.0
dnspython==2.4.2